Moved user settings to avoid cyclic imports in the future
